Fixed a compiler warning on systems where ptrdiff_t is not int.
CVS-ID: wrapper.c 1.8

+++ wrapper.c
@@ -312,7 +312,7 @@ static void *run(void *arg)
 		if (strcmp(crypt_ra(tests[i].pw, tests[i].hash, &data, &size),
 		    tests[i].hash)) {
 			printf("%d: FAILED (crypt_ra/%d/%lu)\n",
-				(char *)arg - (char *)0, i, count);
+				(int)((char *)arg - (char *)0), i, count);
 			free(data);
 			return NULL;
 		}
